matcher: move test-only MatchFromStrings into export_test.go
MatchFromStrings had no production callers, only matcher tests. Relocating it into an export_test.go file keeps it out of the shipped binary while preserving the test coverage of Match.DebugString.
